backend/requirements.txt
python-dotenv==1.0.1
supabase==2.11.0
web3==7.6.0
eth_abi>=5.0.0
openai==1.59.9
apscheduler==3.10.4
requests==2.32.3
